get config and paths in umi-serve
