+++ packages/umi-build-dev/src/getPaths.js
@@ -1,9 +1,5 @@
 import { join } from 'path';
-import { existsSync, statSync } from 'fs';
-
-function test(path) {
-  return existsSync(path) && statSync(path).isDirectory();
-}
+import getPaths from 'umi-core/lib/getPaths';
 
 function template(path) {
   return join(__dirname, '../template', path);
@@ -11,47 +7,8 @@ function template(path) {
 
 export default function(service) {
   const { cwd, config } = service;
-  const outputPath = config.outputPath || './dist';
-
-  let pagesPath = 'pages';
-  if (process.env.PAGES_PATH) {
-    pagesPath = process.env.PAGES_PATH;
-  } else {
-    if (test(join(cwd, 'src/page'))) {
-      pagesPath = 'src/page';
-    }
-    if (test(join(cwd, 'src/pages'))) {
-      pagesPath = 'src/pages';
-    }
-    if (test(join(cwd, 'page'))) {
-      pagesPath = 'page';
-    }
-  }
-
-  const absPagesPath = join(cwd, pagesPath);
-  const absSrcPath = join(absPagesPath, '../');
-
-  const envAffix = process.env.NODE_ENV === 'development' ? '' : `-production`;
-  const tmpDirPath = process.env.UMI_TEMP_DIR
-    ? `${process.env.UMI_TEMP_DIR}${envAffix}`
-    : `${pagesPath}/.umi${envAffix}`;
-
-  const absTmpDirPath = join(cwd, tmpDirPath);
-
   return {
-    cwd,
-    outputPath,
-    absOutputPath: join(cwd, outputPath),
-    absNodeModulesPath: join(cwd, 'node_modules'),
-    pagesPath,
-    absPagesPath,
-    absSrcPath,
-    tmpDirPath,
-    absTmpDirPath,
-    absRouterJSPath: join(absTmpDirPath, 'router.js'),
-    absLibraryJSPath: join(absTmpDirPath, 'umi.js'),
-    absRegisterSWJSPath: join(absTmpDirPath, 'registerServiceWorker.js'),
-    absPageDocumentPath: join(absPagesPath, 'document.ejs'),
+    ...getPaths({ cwd, config }),
     defaultEntryTplPath: template('entry.js.tpl'),
     defaultRouterTplPath: template('router.js.tpl'),
     defaultHistoryTplPath: template('history.js.tpl'),
