+++ packages/umi-core/package.json
@@ -1,7 +1,6 @@
 {
   "name": "umi-core",
   "version": "0.1.0",
-  "main": "lib/index.js",
   "dependencies": {
     "extend2": "1.0.0"
   },
