--- packages/umi-core/src/index.js
@@ -1,7 +1,0 @@
-export {
-  default as getUserConfig,
-  getConfigByConfigFile,
-  getConfigFile,
-  cleanConfigRequireCache,
-  getConfigPaths,
-} from './getUserConfig';
