+++ packages/umi-serve/package.json
@@ -14,6 +14,7 @@
     "chalk": "2.4.1",
     "express": "4.16.4",
     "umi-mock": "0.1.0",
+    "umi-core": "0.1.0",
     "umi-utils": "1.3.1",
     "serve-static": "1.13.2",
     "yargs-parser": "11.1.1"
